fix(web): Compute frame content boxes on the store so reactivity works
